# docs/swagger.yaml
definitions:
...
  models.ConsultationVisit:
    properties:
      fecha_hora_visita:
        type: string
      hospitalEmployee:
        $ref: '#/definitions/models.HospitalEmployee'
      id_paciente:
        type: integer
      id_personal_hospital:
        type: integer
      notas_medicas:
        type: string
      patient:
        $ref: '#/definitions/models.Patient'
      razon_visita:
        type: string
      resultados_examenes:
        type: string
    type: object
  models.DemographicData:
    properties:
      estado_civil:
        type: string
      informacion_etnica_racial:
        type: string
      informacion_laboral:
        type: string
      nacionalidad:
        type: string
    type: object
  models.DiagnosticImage:
    properties:
      consultationVisit:
        $ref: '#/definitions/models.ConsultationVisit'
      descripcion:
        type: string
      fecha_imagen:
        type: string
      id_consulta:
        type: integer
      interpretacion_imagen:
        type: string
      ruta_archivo_externo:
        type: string
      tipo_imagen:
        type: string
    type: object
  models.HospitalEmployee:
    properties:
      direccion:
        type: string
      fecha_nacimiento:
        type: string
      genero:
        type: string
      id_identificacion:
        type: integer
      id_rol:
        description: Llave foránea que referencia la tabla roles
        type: integer
      identification:
        $ref: '#/definitions/models.Identification'
      nombre_completo:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        description: Carga la información del rol relacionado
      telefono:
        type: string
    type: object
  models.Identification:
    properties:
      numero_documento:
        type: string
      tipo_documento:
        type: string
    type: object
...
  models.MedicalRecord:
    properties:
      alergias:
        type: string
      antecedentes_familiares:
        type: string
      diagnosticos_previos:
        type: string
      enfermedades_cronicas:
        type: string
      fecha_actualizacion_historial:
        type: string
      fecha_creacion_historial:
        type: string
      historial_cirugias:
        type: string
      id_paciente:
        description: Llave foránea que referencia la tabla roles
        type: integer
      medicamentos_actuales:
        type: string
      patient:
        $ref: '#/definitions/models.Patient'
    type: object
  models.Patient:
    properties:
      demographicData:
        $ref: '#/definitions/models.DemographicData'
      direccion:
        type: string
      fecha_nacimiento:
        type: string
      genero:
        type: string
      id_datos_demograficos:
        type: integer
      id_identificacion:
        type: integer
      id_rol:
        description: Llave foránea que referencia la tabla roles
        type: integer
      id_usuarios:
        type: integer
      identification:
        $ref: '#/definitions/models.Identification'
      nombre_completo:
        type: string
      numero_seguro_social:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        description: Carga la información del rol relacionado
      telefono:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.Role:
    properties:
      descripcion:
        type: string
      nombre_rol:
        type: string
    type: object
...
  models.User:
    properties:
      contraseña:
        type: string
      id_rol:
        description: Llave foránea que referencia la tabla roles
        type: integer
      nombre_usuario:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        description: Carga la información del rol relacionado
    type: object
...
  request.CreateDiagnosticImageDTO:
    properties:
      descripcion:
        type: string
      fecha_imagen:
        example: "2025-01-20"
        type: string
      id_consulta:
        type: integer
      interpretacion_imagen:
        type: string
      ruta_archivo_externo:
        type: string
      tipo_imagen:
        type: string
    required:
    - descripcion
    - fecha_imagen
    - id_consulta
    - interpretacion_imagen
    - ruta_archivo_externo
    - tipo_imagen
    type: object
...
  request.CreateMedicalRecordDTO:
    properties:
      alergias:
        type: string
      antecedentes_familiares:
        type: string
      diagnosticos_previos:
        type: string
      enfermedades_cronicas:
        type: string
      fecha_actualizacion_historial:
        type: string
      fecha_creacion_historial:
        type: string
      historial_cirugias:
        type: string
      id_paciente:
        description: ID del rol asociado, requerido
        type: integer
      medicamentos_actuales:
        type: string
    required:
    - alergias
    - antecedentes_familiares
    - diagnosticos_previos
    - enfermedades_cronicas
    - fecha_actualizacion_historial
    - fecha_creacion_historial
    - historial_cirugias
    - id_paciente
    - medicamentos_actuales
    type: object
...
info:
  contact: {}
paths:
...
  /diagnostic_images:
    get:
      description: Retrieve a list of all diagnostic_images in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DiagnosticImage'
            type: array
      summary: List all diagnostic_images
      tags:
      - diagnostic_images
    post:
      consumes:
      - application/json
      description: Adds a new diagnostic_image to the system
      parameters:
      - description: DiagnosticImage data
        in: body
        name: diagnostic_image
        required: true
        schema:
          $ref: '#/definitions/request.CreateDiagnosticImageDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.DiagnosticImage'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new diagnostic_image
      tags:
      - diagnostic_images
  /diagnostic_images/{id}:
    delete:
      description: Remove a diagnostic_image by its ID
      parameters:
      - description: DiagnosticImage ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a diagnostic_image
      tags:
      - diagnostic_images
    get:
      description: Retrieve a single diagnostic_image by its ID
      parameters:
      - description: DiagnosticImage ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DiagnosticImage'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a diagnostic_image by ID
      tags:
      - diagnostic_images
    put:
      consumes:
      - application/json
      description: Update the information of an existing diagnostic_image
      parameters:
      - description: DiagnosticImage ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated diagnostic_image data
        in: body
        name: diagnostic_image
        required: true
        schema:
          $ref: '#/definitions/request.CreateDiagnosticImageDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DiagnosticImage'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a diagnostic_image
      tags:
      - diagnostic_images
...
  /medical_records:
    get:
      description: Retrieve a list of all medical_records in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MedicalRecord'
            type: array
      summary: List all medical_records
      tags:
      - medical_records
    post:
      consumes:
      - application/json
      description: Adds a new medical_record to the system
      parameters:
      - description: MedicalRecord data
        in: body
        name: medical_record
        required: true
        schema:
          $ref: '#/definitions/request.CreateMedicalRecordDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.MedicalRecord'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new medical_record
      tags:
      - medical_records
  /medical_records/{id}:
    delete:
      description: Remove a medical_record by its ID
      parameters:
      - description: MedicalRecord ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a medical_record
      tags:
      - medical_records
    get:
      description: Retrieve a single medical_record by its ID
      parameters:
      - description: MedicalRecord ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MedicalRecord'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a medical_record by ID
      tags:
      - medical_records
    put:
      consumes:
      - application/json
      description: Update the information of an existing medical_record
      parameters:
      - description: MedicalRecord ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated medical_record data
        in: body
        name: medical_record
        required: true
        schema:
          $ref: '#/definitions/request.CreateMedicalRecordDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MedicalRecord'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a medical_record
      tags:
      - medical_records
...
swagger: "2.0"